chore(NODE-3717): test reorg - the conclusion (#3105)
