+++ package.json
@@ -104,8 +104,8 @@
     "check:eslint": "eslint -v && eslint --max-warnings=0 --ext '.js,.ts' src test",
     "check:tsd": "tsd --version && tsd",
     "check:dts": "./node_modules/typescript/bin/tsc --noEmit mongodb.d.ts && tsd",
-    "check:test": "mocha --file test/tools/runner --recursive test/functional test/integration",
-    "check:unit": "mocha --recursive test/unit/",
+    "check:test": "mocha --file test/tools/runner test/integration",
+    "check:unit": "mocha test/unit/",
     "check:ts": "./node_modules/typescript/bin/tsc -v && ./node_modules/typescript/bin/tsc --noEmit",
     "check:atlas": "mocha --config \"test/manual/mocharc.json\" test/manual/atlas_connectivity.test.js",
     "check:adl": "mocha --file test/tools/runner test/manual/data_lake.test.js",
@@ -116,7 +116,7 @@
     "check:ldap": "mocha --config \"test/manual/mocharc.json\" test/manual/ldap.test.js",
     "check:socks5": "mocha --config \"test/manual/mocharc.json\" test/manual/socks5.test.ts",
     "check:csfle": "mocha --file test/tools/runner test/integration/client-side-encryption",
-    "check:snappy": "mocha --file test/tools/runner test/functional/unit_snappy.test.js",
+    "check:snappy": "mocha test/unit/assorted/snappy.test.js",
     "prepare": "node etc/prepare.js",
     "release": "standard-version -i HISTORY.md",
     "test": "npm run check:lint && npm run test:all",
