--- test/functional/core/key/keyfile.key
@@ -1,1 +1,0 @@
-THIS IS A SECRET KEYFILE FOR REPLICA SETS BWAHAHAHAH
