+++ test/integration/auth/mongodb_aws.test.js
@@ -2,9 +2,10 @@
 const { expect } = require('chai');
 
 describe('MONGODB-AWS', function () {
-  before(function () {
+  beforeEach(function () {
     const MONGODB_URI = process.env.MONGODB_URI;
     if (!MONGODB_URI || MONGODB_URI.indexOf('MONGODB-AWS') === -1) {
+      this.currentTest.skipReason = 'requires MONGODB_URI to contain MONGODB-AWS auth mechanism';
       this.skip();
     }
   });
