+++ test/integration/node-specific/examples/readme.md
@@ -1,0 +1,5 @@
+# Node.js Driver Usage Examples
+
+This directory contains examples of how to use the various Node.js driver features, which are written in test files to ensure that the sample code actually compiles.
+
+**NOTE:** The examples in this folder are used in the [docs build](https://github.com/mongodb/docs-worker-pool/blob/6a8c5843f50236d39019d5cae341437149ffa6ab/makefiles/Makefile.docs-mongodb-internal#L95-L109) via a direct curl from the repository. Any change in the location of this folder or its contents **must** be coordinated with the docs team.
