+++ test/tools/runner/chai-addons.js
@@ -4,6 +4,6 @@
 const chai = require('chai');
 chai.use(require('sinon-chai'));
 chai.use(require('chai-subset'));
-chai.use(require('../../functional/spec-runner/matcher').default);
+chai.use(require('../spec-runner/matcher').default);
 
 chai.config.truncateThreshold = 0;
