+++ test/unit/assorted/server_discovery_and_monitoring.spec.test.js
@@ -16,12 +16,7 @@ const {
   MongoServerError
 } = require('../../../src/error');
 const { eachAsyncSeries, ns } = require('../../../src/utils');
-
-const chai = require('chai');
-chai.use(require('chai-subset'));
-chai.use(require('../../functional/spec-runner/matcher').default);
-
-const expect = chai.expect;
+const { expect } = require('chai');
 
 const specDir = path.resolve(__dirname, '../../spec/server-discovery-and-monitoring');
 function collectTests() {
